requirements.txt
Flask==2.2.2
openai==0.25.0
python-dotenv==0.21.0
